+++ src/txn_impl.cpp
@@ -10,7 +10,7 @@ namespace calicodb
 {
 
 TxnImpl::TxnImpl(Pager &pager, Status &status, bool write)
-    : m_schema(pager, status, write),
+    : m_schema(pager, status),
       m_pager(&pager),
       m_status(&status),
       m_write(write)
@@ -33,15 +33,30 @@ auto TxnImpl::status() const -> Status
 auto TxnImpl::new_table(const TableOptions &options, const std::string &name, Table *&out) -> Status
 {
     out = nullptr;
-    CALICODB_TRY(*m_status);
-    return m_schema.new_table(options, name, out);
+    auto s = *m_status;
+    if (s.is_ok()) {
+        auto altered = options;
+        if (altered.create_if_missing) {
+            altered.create_if_missing = m_write;
+        }
+        if ((s = m_schema.new_table(altered, name, out)).is_ok()) {
+            table_impl(out)->m_readonly = !m_write;
+        }
+    }
+    return s;
 }
 
 auto TxnImpl::drop_table(const std::string &name) -> Status
 {
-    CALICODB_TRY(*m_status);
-    // Schema object disallows dropping tables during readonly transactions.
-    return m_schema.drop_table(name);
+    if (!m_write) {
+        return Status::readonly();
+    }
+    auto s = *m_status;
+    if (s.is_ok()) {
+        // Schema class disallows dropping tables during readonly transactions.
+        s = m_schema.drop_table(name);
+    }
+    return s;
 }
 
 auto TxnImpl::commit() -> Status
@@ -49,8 +64,11 @@ auto TxnImpl::commit() -> Status
     if (!m_write) {
         return Status::ok();
     }
-    CALICODB_TRY(*m_status);
-    return m_pager->commit();
+    auto s = *m_status;
+    if (s.is_ok()) {
+        s = m_pager->commit();
+    }
+    return s;
 }
 
 auto TxnImpl::rollback() -> void
@@ -64,11 +82,15 @@ auto TxnImpl::rollback() -> void
 auto TxnImpl::vacuum() -> Status
 {
     if (!m_write) {
-        return readonly_transaction();
+        return Status::readonly();
     }
-    CALICODB_TRY(*m_status);
-    m_pager->set_status(vacuum_freelist());
-    return *m_status;
+    auto s = *m_status;
+    if (s.is_ok()) {
+        if (!(s = vacuum_freelist()).is_ok()) {
+            m_pager->set_status(s);
+        }
+    }
+    return s;
 }
 
 auto TxnImpl::vacuum_freelist() -> Status
@@ -87,8 +109,11 @@ auto TxnImpl::vacuum_freelist() -> Status
         return Status::ok();
     }
 
-    m_pager->set_page_count(pgid.value);
-    return m_schema.vacuum_finish();
+    auto s = m_schema.vacuum_finish();
+    if (s.is_ok()) {
+        m_pager->set_page_count(pgid.value);
+    }
+    return s;
 }
 
 auto TxnImpl::TEST_validate() const -> void
